refactor(rust/engine): adopt generated context-planning contracts (Slice A.2)
Replace the engine-local ContextRequest / ContextCandidate / ContextDecision /
ModelInvocationContextSnapshot twins with the generated cross-runtime pipeline
types (crate::model), re-exported under their engine names. The snapshot's
portability/delegated_state are now carried through the nested
InvocationContextState, and int fields adopt the canonical i32 width; the engine
loop keeps usize internally and casts only at the ContextRequest boundary.

validate()/validate_for() move to same-crate inherent impls on the generated
ModelInvocationContextSnapshot. The now-identity generated_snapshot/
generated_decision/generated_request bridge functions are deleted; the ModelPort
adapter clones the snapshot straight across the provider boundary.

Replay/harness normalization is insensitive to the snapshot payload shape, so the
flat->nested wire change leaves spec/vectors/harness replay vectors green.

Gate: model 419/0, clippy --workspace --lib -D warnings clean, full workspace
all-targets pass (incl. replay/harness/resume). Net -89 lines.

Copilot-Session: 6aff469d-9c10-44ec-b343-c32e0a44af7f

+++ runtime/rust/prompty/src/harness.rs
@@ -464,7 +464,7 @@ impl ModelPort for ReferenceModelPort {
         let response = (self.callback)(TurnModelRequest {
             session_id: request.context.session_id.clone(),
             turn_id: request.context.turn_id.clone(),
-            iteration: request.context.iteration as i32,
+            iteration: request.context.iteration,
             inputs: self.inputs.clone(),
             options: Some(self.options.clone()),
             tool_results,
